Support for authentication path other than /auth on site route
Add .mountPoint to your request and it will be added tot the
authentication callback path (which is called by the remote
authentication service). This is handy when the app is not mounted on
the root of the server, for example when you host multiple node.js apps
on a single http port using a router like node-http-proxy.

Just twitter and the oauth2 based services have been updated.